isolating the databaseclient interface and placing it in a separate interface.go to make it modular
